requirements/base.txt
# Wheel 0.25+ needed to install certain packages on CPython 3.5+
# like Pillow and psycopg2
# See http://bitly.com/wheel-building-fails-CPython-35
# Verified bug on Python 3.5.1
wheel==0.29.0


# Bleeding edge Django
django==1.10.7

# Configuration
django-environ==0.4.3

# Models
django-model-utils==3.0.0

# Images
Pillow==4.1.1

# Password storage
argon2-cffi==16.3.0

# Python-PostgreSQL Database Adapter
psycopg2==2.7.1


# Time zones support
pytz==2017.2

# Redis support
django-redis==4.8.0
redis>=2.10.5


celery==3.1.25


djangorestframework==3.6.3
